+++ src/main/java/org/example/topprogramingstoriesdashboard/scraper/TopStoriesGateway.java
@@ -3,5 +3,5 @@ package org.example.topprogramingstoriesdashboard.scraper;
 import java.util.List;
 
 public interface TopStoriesGateway {
-    List<Integer> getTopStories();
+    List<Long> getTopStories();
 }
